edit UserProfile weight to int, edit UserProfileController´(deleted all methods) added addUserProfile method, edit UserProfileService addUserProfile method, added ControllerExceptionHandler but not working yet :)

+++ Spring/src/main/java/de/sep/calocalendar/entities/UserProfile.java
@@ -20,7 +20,7 @@ public class UserProfile {
     private String userName;
     private int gender;
     private int age;
-    private double weight;
+    private int weight;
     private int levelOfPhysicalActivity;
 
 }
